Revert "bit of refactoring"
This reverts commit bf0d696f5b83f88b356361965b17c95c0ab3f633.
